Unify dump custom/gz and custom/zstd tests
